rename server/db/couchdb/list.js -> server/db/couchdb/databases.js
to avoid confusion with listings
