+++ scripts/backup_databases/lib/get_databases_names.js
@@ -1,4 +1,4 @@
-const dbsBaseNames = Object.keys(require('db/couchdb/list'))
+const dbsBaseNames = Object.keys(require('db/couchdb/databases'))
 
 module.exports = suffix => {
   if (suffix) return dbsBaseNames.map(dbBaseName => `${dbBaseName}-${suffix}`)
