+++ scripts/replicate_tests_db.js
@@ -7,7 +7,7 @@ const error_ = require('lib/error/error')
 const dbHost = require('config').db.getOrigin()
 
 const dbUrl = dbName => `${dbHost}/${dbName}`
-const dbsBaseNames = Object.keys(require('db/couchdb/list'))
+const dbsBaseNames = Object.keys(require('db/couchdb/databases'))
 
 const replicate = async dbName => {
   const dbTestName = `${dbName}-tests`
