+++ server/db/couchdb/base.js
@@ -2,13 +2,13 @@ const CONFIG = require('config')
 const _ = require('builders/utils')
 const getDbApi = require('./cot_base')
 const memoize = require('lib/utils/memoize')
-const list = require('./list')
+const databases = require('db/couchdb/databases')
 
 module.exports = (dbBaseName, designDocName) => {
   const dbName = CONFIG.db.name(dbBaseName)
   // If no designDocName is provided while there are defined design docs for this database,
   // assumes that it is the default design doc, which has the same name as the dbBaseName
-  if (list[dbBaseName].length > 0 && designDocName == null) {
+  if (databases[dbBaseName].length > 0 && designDocName == null) {
     designDocName = dbBaseName
   }
   return getHandler(dbBaseName, dbName, designDocName)
@@ -23,13 +23,13 @@ const getHandler = memoize((dbBaseName, dbName, designDocName) => {
 
 // Not using error_ as that would make hard to solve cirucular dependencies
 const validate = (dbBaseName, designDocName) => {
-  if (!list[dbBaseName]) {
+  if (!databases[dbBaseName]) {
     throw new Error(`unknown dbBaseName: ${dbBaseName}`)
   }
 
   const jsDesignDocName = `${designDocName}.js`
 
-  if (designDocName && !(list[dbBaseName].includes(designDocName) || list[dbBaseName].includes(jsDesignDocName))) {
+  if (designDocName && !(databases[dbBaseName].includes(designDocName) || databases[dbBaseName].includes(jsDesignDocName))) {
     throw new Error(`unknown designDocName: ${designDocName}`)
   }
 }
